test(pg): the promote-lane-shared script test requires the test database instead of skipping (#878)
The file read OPENBRAIN_TEST_DATABASE_URL itself and swapped in describe.skip
when it was unset, so an absent database reported `0 pass, 10 skip, 0 fail` and
exited 0 -- indistinguishable at the exit code from a suite that ran and passed.
The Postgres write-path behavior it covers could drift with CI staying green.

It now calls requireTestDatabaseUrl() at module scope, so an absent database
throws test_database_required and takes the run down loudly.

The 619-line file is split by subject at the same time:

- scripts/promote-lane-shared.test.ts keeps the cursor-stall suite, name
  unchanged.
- scripts/promote-lane-shared-clustering.test.ts holds the clustering and
  cursor-loop tests as "runSharedPromoter clustering and cursor loops
  (live Postgres)".
- scripts/test-support/promote-lane-shared-helpers.ts holds the seeding,
  cleanup, and environment plumbing both suites drive as
  createPromoterHarness(pool, databaseUrl). It holds no test and creates no
  pool: each test file owns its module-scope pool and hands it in.

Test bodies, names, order, and assertions move verbatim. The oxlint findings the
touched files already carried are cleared without disable comments: 23 non-null
assertions become expectDefined(value, label) guards exported from the harness,
new Array(n).fill(0) becomes Array.from, and the long describe callbacks come
under 100 code lines by hoisting five test bodies into module-scope functions
called with the same test names.

scripts/assert-db-tests-ran.ts follows in the same commit, counts read from
JUnit rather than tallied by eye: the kept suite emits 1 and the new clustering
suite emits 8, so MIN_TOTAL_LIVE_TESTCASES moves 265 -> 266.

Done-means: scripts/done-means/878-pg-tests-require-database.sh
