+++ scripts/promote-lane-shared.test.ts
@@ -1,214 +1,35 @@
-import {
-  afterAll,
-  afterEach,
-  beforeEach,
-  describe,
-  expect,
-  test,
-} from "bun:test";
-import { mkdtempSync, readFileSync, rmSync } from "node:fs";
+import { afterAll, afterEach, beforeEach, describe, expect, test } from "bun:test";
+import { mkdtempSync, rmSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { Pool } from "pg";
-import { parseArgs, runSharedPromoter } from "./promote-lane-shared.ts";
-import { sharedNamespaceConfig } from "../src/shared-namespace.ts";
+import { runSharedPromoter } from "./promote-lane-shared.ts";
+import { requireTestDatabaseUrl } from "./test-support/require-test-database.ts";
+import { createPromoterHarness } from "./test-support/promote-lane-shared-helpers.ts";
 
-// ── Cursor-stall + dry-run-embedding coverage (Issue #161, hybrid timing) ──
+// ── Cursor-stall coverage (Issue #161, hybrid timing) ──
 //
 // runSharedPromoter creates its OWN pool via createPool(), which reads
 // DB_HOST/DB_USER/DB_NAME/DB_PORT/DB_PASSWORD from the environment (NOT a
-// connection string). So before each DB-gated test we parse
-// OPENBRAIN_TEST_DATABASE_URL into those env vars. A separate, URL-built Pool is
-// used here only for seeding and assertions.
+// connection string). So before each test the harness parses the test database
+// URL into those variables. A separate, URL-built Pool is used here only for
+// seeding and assertions.
 //
-// SME HARD RULE (docs/sme/correctness.md): SQL write-path behavior — cursor
+// SME HARD RULE (docs/sme/correctness.md): SQL write-path behavior -- cursor
 // advancement persisted to the state file, the JOIN/cursor predicates, and the
-// metadata jsonb shape — CANNOT be caught by a mock pool. These tests run the
-// real query through real Postgres and are env-gated so the default suite stays
-// infra-free. Run with:
-//   OPENBRAIN_TEST_DATABASE_URL=postgres://user:pass@host:5432/db \
-//     bun test scripts/promote-lane-shared.test.ts
-const DB_URL = process.env.OPENBRAIN_TEST_DATABASE_URL;
-const dbDescribe = DB_URL ? describe : describe.skip;
+// metadata jsonb shape -- CANNOT be caught by a mock pool. These tests run the
+// real query through real Postgres, and per issue #878 they DEMAND the test
+// database rather than skipping themselves when it is absent.
+const databaseUrl = requireTestDatabaseUrl();
+const pool = new Pool({ connectionString: databaseUrl });
 
-const DEFAULT_MIN = 24;
-// minLen <= len < minLen*1.5 → manual-review. minLen=24 → [24, 36).
-const MANUAL_REVIEW_CONTENT = "x".repeat(30); // len 30, in the ambiguity band.
-// len >= minLen*1.5 (>=36) and a shareable type → share.
-const SHARE_CONTENT =
-  "This is a substantive shared-worthy decision about the schema design choices.";
-
-dbDescribe("runSharedPromoter cursor-stall fix (live Postgres)", () => {
-  // Build a seeding/assertion pool from the URL.
-  const pool = new Pool({ connectionString: DB_URL });
-  const ns = "test-promote-lane-shared-live";
+describe("runSharedPromoter cursor-stall fix (live Postgres)", () => {
+  const h = createPromoterHarness(pool, databaseUrl);
   let tmpDir: string;
   let stateFile: string;
-  let savedEnv: Record<string, string | undefined>;
-
-  // Translate the connection URL into the env vars createPool() expects so the
-  // runner's internal pool targets the same test database.
-  function applyDbEnv(url: string): void {
-    const u = new URL(url);
-    savedEnv = {
-      DB_HOST: process.env.DB_HOST,
-      DB_PORT: process.env.DB_PORT,
-      DB_NAME: process.env.DB_NAME,
-      DB_USER: process.env.DB_USER,
-      DB_PASSWORD: process.env.DB_PASSWORD,
-    };
-    process.env.DB_HOST = u.hostname;
-    process.env.DB_PORT = u.port || "5432";
-    process.env.DB_NAME = u.pathname.replace(/^\//, "") || "open_brain";
-    process.env.DB_USER = decodeURIComponent(u.username);
-    process.env.DB_PASSWORD = decodeURIComponent(u.password);
-  }
-
-  function restoreDbEnv(): void {
-    for (const [k, v] of Object.entries(savedEnv)) {
-      if (v === undefined) delete process.env[k];
-      else process.env[k] = v;
-    }
-  }
-
-  // The runner's events loop inserts promoted lane events into the REAL physical
-  // shared-kb namespace (config.physicalSharedNamespace), not ns+"-shared", so we
-  // resolve it the same way the runner does for an exhaustive cleanup.
-  const sharedPhysicalNs = sharedNamespaceConfig().physicalSharedNamespace;
-
-  async function cleanupNs(): Promise<void> {
-    // Make cleanup namespace-exhaustive across EVERYTHING any test in this file
-    // can create, deleting by namespace rather than by fragile source strings.
-    //
-    // Sources of residue per test:
-    //  1. Seeded source thoughts  → namespace = ns
-    //  2. promoteEntry-promoted thought/decision COPIES → namespace = ns+"-shared".
-    //     These keep the SOURCE row's `source` value (NOT 'lane-shared-promotion')
-    //     so a source-string predicate misses them entirely.
-    //  3. Events + lanes → ob_session_events cascade-delete with ob_session_lanes.
-    //  4. Promoted lane-event copies → namespace = sharedPhysicalNs (real shared-kb),
-    //     source = 'lane-shared-promotion', provenance source_physical_namespace = ns.
-    //
-    // ob_links created by thought-cluster supplementation (#173) live in the
-    // pinned shared namespace (ns+"-shared"); also sweep the real shared-kb ns
-    // for any auto-cluster links the events loop could have produced.
-    await pool.query(
-      `DELETE FROM ob_links
-        WHERE relation = 'supplements'
-          AND namespace = ANY($1::text[])
-          AND metadata->>'clustered_by' = 'lane-shared-promoter'`,
-      [[ns + "-shared", sharedPhysicalNs]],
-    );
-    // Lanes/events: delete lanes for the test ns (events cascade).
-    await pool.query("DELETE FROM ob_session_lanes WHERE namespace = $1", [ns]);
-    // Seeded source rows AND promoteEntry-promoted copies, by namespace.
-    await pool.query(
-      "DELETE FROM thoughts WHERE namespace = ANY($1::text[])",
-      [[ns, ns + "-shared"]],
-    );
-    await pool.query(
-      "DELETE FROM decisions WHERE namespace = ANY($1::text[])",
-      [[ns, ns + "-shared"]],
-    );
-    // Promoted lane-event copies that land in the real shared-kb namespace are
-    // scoped to THIS test's provenance so we never touch unrelated shared-kb rows.
-    await pool.query(
-      `DELETE FROM thoughts
-       WHERE namespace = $1
-         AND source = 'lane-shared-promotion'
-         AND promoted_from->>'source_physical_namespace' = $2`,
-      [sharedPhysicalNs, ns],
-    );
-    // Belt-and-suspenders: any promoted copy whose provenance target points at
-    // this test's shared namespace, regardless of which physical ns it landed in.
-    await pool.query(
-      `DELETE FROM thoughts
-       WHERE promoted_from->>'source_physical_namespace' = $1`,
-      [ns],
-    );
-  }
-
-  function explicitNominationMetadata(): Record<string, unknown> {
-    return {
-      share_candidate: true,
-      memory_lifecycle_action: "nominate_shared",
-    };
-  }
-
-  async function seedLane(): Promise<string> {
-    const { rows } = await pool.query(
-      `INSERT INTO ob_session_lanes (session_key, namespace, status, created_by)
-       VALUES ($1, $2, 'active', $3)
-       RETURNING id`,
-      ["promote-test-lane", ns, "test"],
-    );
-    return rows[0].id as string;
-  }
-
-  /** Seed an event. created_at is explicit so cursor ordering is deterministic. */
-  async function seedEvent(
-    laneId: string,
-    content: string,
-    createdAt: string,
-  ): Promise<string> {
-    const { rows } = await pool.query(
-      `INSERT INTO ob_session_events
-         (lane_id, event_type, content, importance, metadata, content_hash, created_by, created_at)
-       VALUES ($1, 'fact', $2, 'warm', $3::jsonb, $4, 'test', $5::timestamptz)
-       RETURNING id`,
-      [
-        laneId,
-        content,
-        JSON.stringify(explicitNominationMetadata()),
-        // unique-ish hash so ON CONFLICT (lane_id, content_hash) won't collide
-        `hash-${content.length}-${createdAt}`,
-        createdAt,
-      ],
-    );
-    return rows[0].id as string;
-  }
-
-  async function seedThought(
-    content: string,
-    createdAt: string,
-  ): Promise<string> {
-    const { rows } = await pool.query(
-      `INSERT INTO thoughts
-         (content, namespace, extracted_metadata, created_by, created_at)
-       VALUES ($1, $2, $3::jsonb, 'test', $4::timestamptz)
-       RETURNING id`,
-      [content, ns, JSON.stringify(explicitNominationMetadata()), createdAt],
-    );
-    return rows[0].id as string;
-  }
-
-  function makeArgs(apply: boolean): ReturnType<typeof parseArgs> {
-    const base = parseArgs([
-      "--state-file",
-      stateFile,
-      "--min-content-length",
-      String(DEFAULT_MIN),
-      "--batch-size",
-      "50",
-      "--max-apply",
-      "50",
-      "--delay-ms",
-      "0",
-    ]);
-    base.apply = apply;
-    // Pin the target namespace away from real shared-kb to avoid cross-talk.
-    base.targetNamespace = ns + "-shared";
-    return base;
-  }
-
-  function readState(): {
-    cursors: Record<string, { created_at?: string; id?: string }>;
-  } {
-    return JSON.parse(readFileSync(stateFile, "utf8"));
-  }
 
   test("candidate-only rows are not shared-kb nominations", async () => {
-    const laneId = await seedLane();
+    const laneId = await h.seedLane();
     await pool.query(
       `INSERT INTO ob_session_events
          (lane_id, event_type, content, importance, metadata, content_hash, created_by, created_at)
@@ -232,7 +53,7 @@ dbDescribe("runSharedPromoter cursor-stall fix (live Postgres)", () => {
        VALUES ($1, $2, $3::jsonb, 'test', $4::timestamptz)`,
       [
         "Candidate-only thought should not be promoted by presence alone.",
-        ns,
+        h.ns,
         JSON.stringify({
           share_candidate: true,
           memory_lifecycle_action: "candidate",
@@ -243,377 +64,36 @@ dbDescribe("runSharedPromoter cursor-stall fix (live Postgres)", () => {
       ],
     );
 
-    const receipt = await runSharedPromoter(makeArgs(true));
+    const receipt = await runSharedPromoter(h.makeArgs(true, stateFile));
     expect(receipt.scanned).toBe(0);
 
     const { rows: sharedCopies } = await pool.query(
       `SELECT id FROM thoughts
         WHERE namespace = $1
           AND content = 'Candidate-only thought should not be promoted by presence alone.'`,
-      [ns + "-shared"],
+      [h.ns + "-shared"],
     );
     expect(sharedCopies.length).toBe(0);
   });
 
   beforeEach(async () => {
-    applyDbEnv(DB_URL as string);
+    h.applyDbEnv();
     // Defend against a PRIOR crashed test leaving residue that a namespace-wide
     // scan (nominatedTableRows has no namespace filter) would re-pick up.
-    await cleanupNs();
+    await h.cleanupNs();
     // DEV_TMP (macOS dev) when set; otherwise the OS temp dir so this runs on
     // the Linux CI runner (the Mac /Volumes path does not exist there).
-    tmpDir = mkdtempSync(
-      join(process.env.DEV_TMP ?? tmpdir(), "promote-lane-shared-"),
-    );
+    tmpDir = mkdtempSync(join(process.env.DEV_TMP ?? tmpdir(), "promote-lane-shared-"));
     stateFile = join(tmpDir, "state.json");
   });
 
   afterEach(async () => {
-    await cleanupNs();
-    restoreDbEnv();
+    await h.cleanupNs();
+    h.restoreDbEnv();
     rmSync(tmpDir, { recursive: true, force: true });
   });
 
   afterAll(async () => {
     await pool.end();
-  });
-
-  // ── Thought-cluster supplementation (#173) ──
-  //
-  // We drive the THOUGHTS promoteEntry path with DETERMINISTIC embeddings so the
-  // cosine distance between a promoted thought and a seeded shared-kb anchor is
-  // exact and controllable — no embedding server needed. promoteEntry copies the
-  // source thought's `embedding` column verbatim into the shared-kb copy, so the
-  // promoted row's vector equals the source vector we set here.
-  //
-  // Construction: all vectors are 768-dim unit vectors in the span of basis
-  // components 0 and 1. For unit vectors u=[a,b,0…] and v=[c,d,0…] the cosine
-  // similarity is a*c + b*d, and pgvector's `<=>` (halfvec_cosine_ops) returns
-  // cosine DISTANCE = 1 - similarity. So choosing the dot product picks the band.
-  const EMBED_DIM = 768;
-
-  /** A 768-dim halfvec literal for a unit vector with the given [c0, c1] head. */
-  function unitVec(c0: number, c1: number): string {
-    const rest = new Array(EMBED_DIM - 2).fill(0);
-    return JSON.stringify([c0, c1, ...rest]);
-  }
-
-  // Promoted thought direction: pure basis-0 unit vector.
-  const NEW_VEC = unitVec(1, 0);
-  // similarity 0.85 → distance 0.15 → inside band [0.08, 0.25): supplements.
-  const IN_BAND_VEC = unitVec(0.85, Math.sqrt(1 - 0.85 * 0.85));
-  // similarity 0.5 → distance 0.5 → >= 0.25: out of band, NO link (cluster seed).
-  const FAR_VEC = unitVec(0.5, Math.sqrt(1 - 0.5 * 0.5));
-  // similarity 0.96 → distance 0.04 → < 0.08: exact-dup band, NOT clustered.
-  const DUP_VEC = unitVec(0.96, Math.sqrt(1 - 0.96 * 0.96));
-
-  /** Seed a source thought (namespace=ns) with a nomination flag + embedding. */
-  async function seedThoughtWithEmbedding(
-    content: string,
-    createdAt: string,
-    vecLiteral: string,
-  ): Promise<string> {
-    const { rows } = await pool.query(
-      `INSERT INTO thoughts
-         (content, namespace, extracted_metadata, created_by, created_at,
-          embedding, content_hash)
-       VALUES ($1, $2, $3::jsonb, 'test', $4::timestamptz, $5::halfvec, $6)
-       RETURNING id`,
-      [
-        content,
-        ns,
-        JSON.stringify(explicitNominationMetadata()),
-        createdAt,
-        vecLiteral,
-        `src-hash-${createdAt}`,
-      ],
-    );
-    return rows[0].id as string;
-  }
-
-  /** Seed an EXISTING shared-kb anchor thought with a controlled embedding. */
-  async function seedSharedAnchor(
-    content: string,
-    vecLiteral: string,
-  ): Promise<string> {
-    const { rows } = await pool.query(
-      `INSERT INTO thoughts
-         (content, namespace, created_by, embedding, content_hash)
-       VALUES ($1, $2, 'test', $3::halfvec, $4)
-       RETURNING id`,
-      [content, ns + "-shared", vecLiteral, `anchor-hash-${content.length}-${content}`],
-    );
-    return rows[0].id as string;
-  }
-
-  /** Count supplements links FROM a given new thought. */
-  async function supplementsLinks(
-    fromId: string,
-  ): Promise<Array<{ to_id: string; metadata: Record<string, unknown> }>> {
-    const { rows } = await pool.query(
-      `SELECT to_id, metadata FROM ob_links
-        WHERE from_type = 'thought' AND from_id = $1
-          AND relation = 'supplements' AND archived_at IS NULL`,
-      [fromId],
-    );
-    return rows as Array<{ to_id: string; metadata: Record<string, unknown> }>;
-  }
-
-  test("clustering: in-band neighbour gets a 'supplements' link to the anchor", async () => {
-    const anchorId = await seedSharedAnchor("Existing shared cluster anchor about schema.", IN_BAND_VEC);
-    await seedThoughtWithEmbedding(SHARE_CONTENT, "2026-08-01T00:00:00Z", NEW_VEC);
-
-    const receipt = await runSharedPromoter(makeArgs(true));
-    const thoughts = receipt.sources.thoughts;
-    expect(thoughts).toBeDefined();
-    expect(thoughts!.shared).toBe(1);
-    expect(thoughts!.clustered).toBe(1);
-    expect(receipt.clustered).toBe(1);
-
-    // Find the promoted copy in the shared ns and assert the supplements edge.
-    const { rows: copies } = await pool.query(
-      `SELECT id FROM thoughts
-        WHERE namespace = $1 AND content = $2`,
-      [ns + "-shared", SHARE_CONTENT],
-    );
-    expect(copies.length).toBe(1);
-    const links = await supplementsLinks(copies[0].id);
-    expect(links.length).toBe(1);
-    const link = links[0]!;
-    expect(link.to_id).toBe(anchorId);
-    expect(link.metadata.auto_clustered).toBe(true);
-    expect(Number(link.metadata.distance)).toBeCloseTo(0.15, 2);
-  });
-
-  test("clustering: no in-band neighbour creates NO link (orphan cluster seed)", async () => {
-    await seedSharedAnchor("Distant shared anchor, different topic entirely.", FAR_VEC);
-    await seedThoughtWithEmbedding(SHARE_CONTENT, "2026-08-02T00:00:00Z", NEW_VEC);
-
-    const receipt = await runSharedPromoter(makeArgs(true));
-    expect(receipt.sources.thoughts!.shared).toBe(1);
-    expect(receipt.sources.thoughts!.clustered).toBe(0);
-    expect(receipt.clustered).toBe(0);
-
-    const { rows: copies } = await pool.query(
-      `SELECT id FROM thoughts WHERE namespace = $1 AND content = $2`,
-      [ns + "-shared", SHARE_CONTENT],
-    );
-    const links = await supplementsLinks(copies[0].id);
-    expect(links.length).toBe(0);
-  });
-
-  test("clustering: an exact/near-dup neighbour (dist < 0.08) is NOT clustered", async () => {
-    // The anchor is within the dedup band (distance 0.04). Dedup here is
-    // content_hash-only, so promoteEntry still promotes the new row, but
-    // clustering must SKIP linking because it sits below EXACT_DUP_THRESHOLD.
-    await seedSharedAnchor("Near-duplicate shared anchor close in space.", DUP_VEC);
-    await seedThoughtWithEmbedding(SHARE_CONTENT, "2026-08-03T00:00:00Z", NEW_VEC);
-
-    const receipt = await runSharedPromoter(makeArgs(true));
-    expect(receipt.sources.thoughts!.shared).toBe(1);
-    expect(receipt.sources.thoughts!.clustered).toBe(0);
-
-    const { rows: copies } = await pool.query(
-      `SELECT id FROM thoughts WHERE namespace = $1 AND content = $2`,
-      [ns + "-shared", SHARE_CONTENT],
-    );
-    const links = await supplementsLinks(copies[0].id);
-    expect(links.length).toBe(0);
-  });
-
-  test("clustering: dry-run creates NO link even with an in-band neighbour", async () => {
-    await seedSharedAnchor("Dry-run anchor that would be in band.", IN_BAND_VEC);
-    const srcId = await seedThoughtWithEmbedding(SHARE_CONTENT, "2026-08-04T00:00:00Z", NEW_VEC);
-
-    const receipt = await runSharedPromoter(makeArgs(false));
-    expect(receipt.dry_run).toBe(true);
-    expect(receipt.sources.thoughts!.would_share).toBe(1);
-    expect(receipt.clustered).toBe(0);
-
-    // No promoted copy and therefore no link of any kind was created.
-    const { rows: copies } = await pool.query(
-      `SELECT id FROM thoughts WHERE namespace = $1 AND content = $2`,
-      [ns + "-shared", SHARE_CONTENT],
-    );
-    expect(copies.length).toBe(0);
-    // And no supplements link from the (un-promoted) source either.
-    const links = await supplementsLinks(srcId);
-    expect(links.length).toBe(0);
-  });
-
-  test("events loop: trailing manual-review event does NOT pin the cursor", async () => {
-    const laneId = await seedLane();
-    // share event first (earlier), manual-review event last (trailing).
-    await seedEvent(laneId, SHARE_CONTENT, "2026-01-01T00:00:00Z");
-    const manualId = await seedEvent(
-      laneId,
-      MANUAL_REVIEW_CONTENT,
-      "2026-01-02T00:00:00Z",
-    );
-
-    // First APPLY run: both rows processed, cursor must advance PAST the
-    // trailing manual-review row even though manual-review is not promoted.
-    const first = await runSharedPromoter(makeArgs(true));
-    const eventsReceipt1 = first.sources.ob_session_events;
-    expect(eventsReceipt1).toBeDefined();
-    expect(eventsReceipt1!.scanned).toBe(2);
-    expect(eventsReceipt1!.shared).toBe(1);
-    expect(eventsReceipt1!.manual_review).toBe(1);
-
-    // Cursor now sits on the manual-review row (the last processed row).
-    const cursor1 = readState().cursors.ob_session_events;
-    expect(cursor1?.id).toBe(manualId);
-
-    // Second APPLY run: the manual-review row keeps its share_candidate flag
-    // (only terminal rejects clear it), so without the cursor fix it would be
-    // re-scanned forever. With the fix, the cursor is past it → scanned 0.
-    const second = await runSharedPromoter(makeArgs(true));
-    const eventsReceipt2 = second.sources.ob_session_events;
-    // Either no events source recorded, or scanned excludes the pinned row.
-    expect(eventsReceipt2?.scanned ?? 0).toBe(0);
-  });
-
-  test("events loop: a deterministically-failing event does NOT pin the cursor (poison-pill fix)", async () => {
-    // Poison-pill regression (Issue #161 hardening): in APPLY mode, when a row's
-    // promotion throws inside the try/catch, the loop now advances the cursor
-    // PAST the failed row before `break`-ing. Without that fix the cursor stays
-    // pinned on the failing row, so it (and every row behind it) is re-fetched
-    // on every subsequent run forever.
-    //
-    // FAILURE INJECTION: shareEventToSharedKb does `INSERT INTO thoughts ...`.
-    // The thoughts schema offers no constraint the runner's own write violates
-    // (the only unique index, (content_hash, namespace), is absorbed by the
-    // INSERT's ON CONFLICT DO NOTHING). So we install a temporary BEFORE INSERT
-    // trigger on `thoughts` that RAISEs only when content carries a sentinel
-    // marker. This is a real, deterministic, server-side INSERT failure that the
-    // runner cannot anticipate — exactly the poison-pill shape. The trigger is
-    // test-managed and dropped in this test's finally block.
-    const POISON = `POISON-${Date.now()} `;
-    const poisonContent = POISON + SHARE_CONTENT; // share-worthy, but insert throws.
-
-    await pool.query(`
-      CREATE OR REPLACE FUNCTION test_poison_thought_insert()
-      RETURNS trigger AS $fn$
-      BEGIN
-        IF position('${POISON.trim()}' IN NEW.content) > 0 THEN
-          RAISE EXCEPTION 'poison-pill: deterministic insert failure';
-        END IF;
-        RETURN NEW;
-      END;
-      $fn$ LANGUAGE plpgsql;
-    `);
-    await pool.query(`
-      DROP TRIGGER IF EXISTS test_poison_thought_trg ON thoughts;
-      CREATE TRIGGER test_poison_thought_trg
-        BEFORE INSERT ON thoughts
-        FOR EACH ROW EXECUTE FUNCTION test_poison_thought_insert();
-    `);
-
-    try {
-      const laneId = await seedLane();
-      // First (earlier) event is the poison row that will throw on insert.
-      const poisonId = await seedEvent(
-        laneId,
-        poisonContent,
-        "2026-07-01T00:00:00Z",
-      );
-      // Second (later) event is clean and share-worthy.
-      const cleanId = await seedEvent(
-        laneId,
-        SHARE_CONTENT,
-        "2026-07-02T00:00:00Z",
-      );
-
-      // First APPLY run: poison row is scanned, classified share, then its
-      // shared-kb insert throws. The sweep breaks AFTER advancing the cursor.
-      const first = await runSharedPromoter(makeArgs(true));
-      const events1 = first.sources.ob_session_events;
-      expect(events1).toBeDefined();
-      expect(events1!.scanned).toBe(1); // breaks after the failing row
-      expect(events1!.failed).toBe(1);
-      expect(events1!.shared).toBe(0);
-      // Failure recorded for human follow-up, pinned to the poison row.
-      const failure = first.failures.find((f) => f.id === poisonId);
-      expect(failure).toBeDefined();
-      expect(failure!.source).toBe("ob_session_events");
-
-      // KEY ASSERTION 1: cursor advanced PAST the failed row (not pinned).
-      const cursor1 = readState().cursors.ob_session_events;
-      expect(cursor1?.id).toBe(poisonId);
-
-      // Second APPLY run: forward progress. The poison row's nomination flag is
-      // intentionally left set, but the cursor is now past it, so the runner
-      // does NOT re-fetch it — it reaches the clean row and shares it.
-      const second = await runSharedPromoter(makeArgs(true));
-      const events2 = second.sources.ob_session_events;
-      expect(events2).toBeDefined();
-      // KEY ASSERTION 2: the poison row is NOT re-scanned; only the clean row is.
-      expect(events2!.scanned).toBe(1);
-      expect(events2!.failed).toBe(0);
-      expect(events2!.shared).toBe(1);
-      const cursor2 = readState().cursors.ob_session_events;
-      expect(cursor2?.id).toBe(cleanId);
-    } finally {
-      await pool.query(
-        "DROP TRIGGER IF EXISTS test_poison_thought_trg ON thoughts",
-      );
-      await pool.query(
-        "DROP FUNCTION IF EXISTS test_poison_thought_insert()",
-      );
-    }
-  });
-
-  test("thoughts loop: trailing manual-review thought does NOT pin the cursor", async () => {
-    await seedThought(SHARE_CONTENT, "2026-03-01T00:00:00Z");
-    const manualId = await seedThought(
-      MANUAL_REVIEW_CONTENT,
-      "2026-03-02T00:00:00Z",
-    );
-
-    const first = await runSharedPromoter(makeArgs(true));
-    const thoughtsReceipt1 = first.sources.thoughts;
-    expect(thoughtsReceipt1).toBeDefined();
-    expect(thoughtsReceipt1!.scanned).toBe(2);
-    expect(thoughtsReceipt1!.manual_review).toBe(1);
-
-    const cursor1 = readState().cursors.thoughts;
-    expect(cursor1?.id).toBe(manualId);
-
-    const second = await runSharedPromoter(makeArgs(true));
-    expect(second.sources.thoughts?.scanned ?? 0).toBe(0);
-  });
-
-  test("dry-run does NOT call the embedding endpoint for events", async () => {
-    // The events loop returns would_share BEFORE generateEmbedding when !apply.
-    // To prove the embedding endpoint is not hit, point EMBEDDING_BASE_URL at an
-    // unreachable host: a real generateEmbedding call would error or block on
-    // connect. A clean dry-run with would_share>0 and no failures proves the
-    // embedding call was structurally skipped (the only network call in this
-    // loop is the embedding endpoint). DB connectivity is unaffected because the
-    // runner's pool uses DB_HOST, not EMBEDDING_BASE_URL.
-    const laneId = await seedLane();
-    await seedEvent(laneId, SHARE_CONTENT, "2026-05-01T00:00:00Z");
-
-    const savedEmbedUrl = process.env.EMBEDDING_BASE_URL;
-    // 192.0.2.0/24 is TEST-NET-1 (RFC 5737) — guaranteed non-routable.
-    process.env.EMBEDDING_BASE_URL = "http://192.0.2.1:1/v1";
-    try {
-      const receipt = await runSharedPromoter(makeArgs(false));
-      expect(receipt.dry_run).toBe(true);
-      const events = receipt.sources.ob_session_events;
-      expect(events).toBeDefined();
-      expect(events!.would_share).toBe(1);
-      // No write and no embedding-driven failure occurred.
-      expect(events!.shared).toBe(0);
-      expect(events!.failed).toBe(0);
-      expect(receipt.failures.length).toBe(0);
-    } finally {
-      if (savedEmbedUrl === undefined) delete process.env.EMBEDDING_BASE_URL;
-      else process.env.EMBEDDING_BASE_URL = savedEmbedUrl;
-    }
-
-    // Dry-run must NOT advance the persistent cursor (it only counts).
-    const cursor = readState().cursors.ob_session_events;
-    expect(cursor?.id).toBeUndefined();
   });
 });
